+++ jmix-localized-string-datatype/src/test/java/com/glebfox/jmix/locstr/entity/LocalizedStringTestEntity.java
@@ -1,0 +1,72 @@
+/*
+ * Copyright 2024 Gleb Gorelov.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.glebfox.jmix.locstr.entity;
+
+import com.glebfox.jmix.locstr.datatype.LocalizedString;
+import io.jmix.core.entity.annotation.JmixGeneratedValue;
+import io.jmix.core.metamodel.annotation.InstanceName;
+import io.jmix.core.metamodel.annotation.JmixEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import jakarta.persistence.Version;
+
+import java.util.UUID;
+
+@JmixEntity
+@Table(name = "LOCSTR_TEST_ENTITY")
+@Entity(name = "locstr_LocalizedStringTestEntity")
+public class LocalizedStringTestEntity {
+
+    @JmixGeneratedValue
+    @Column(name = "ID", nullable = false)
+    @Id
+    private UUID id;
+
+    @Version
+    @Column(name = "VERSION", nullable = false)
+    private Integer version;
+
+    @InstanceName
+    @Column(name = "NAME", nullable = false)
+    private LocalizedString name;
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(UUID id) {
+        this.id = id;
+    }
+
+    public Integer getVersion() {
+        return version;
+    }
+
+    public void setVersion(Integer version) {
+        this.version = version;
+    }
+
+    public LocalizedString getName() {
+        return name;
+    }
+
+    public void setName(LocalizedString name) {
+        this.name = name;
+    }
+}